requirements.txt
fastapi~=0.95.0
pydantic~=1.10.7
PyYAML~=6.0
uvicorn~=0.15.0